requirements.txt
Flask==2.2.2
gunicorn
Werkzeug==2.2.2
azure.storage.blob
azure-cosmos
azure.identity
pandas
pycryptodo
azure-keyvault-secrets
azure-keyvault-keys